requirements.txt
Flask==2.0.2
mysql_connector_repackaged==0.3.1
pyserial==3.5
MarkupSafe~=2.0.1
